modified int to long

+++ A big sum.py
@@ -3,5 +3,5 @@ No_input = raw_input()
 input = raw_input().split(' ')
 sum = 0
 for n in input:
-	sum = sum + int(n)
+	sum = sum + long(n)
 print sum
